+++ PndMCMatch/PndMCObject.h
@@ -78,7 +78,7 @@ public:
 
 	friend std::ostream& operator<< (std::ostream& out, const PndMCObject& obj){
 		std::vector<PndMCEntry> stages = obj.GetEntryVector();
-		for (int i = 0; i < stages.size(); i++){
+		for (size_t i = 0; i < stages.size(); i++){
 			if (stages[i].GetNLinks() > 0){
 				out << i << ": ";
 				stages[i].Print(out);
